+++ 3.Software/HelloFinger-rebuild/HelloFinger/interface.cpp
@@ -62,7 +62,7 @@ Interface::Interface(QObject *parent) : QThread(parent)
         HID_Send_Shortcut(usb_handle,(Finger_Type_e)fingertype,func,key,key_len,index);
     });
 
-    connect(mainwindow,&Form_MainWindow::Signal_SetActionFunc,this,[=](uint8_t func,uint32_t action){
+    connect(mainwindow,&Form_MainWindow::Signal_SetActionFunc,this,[=](uint8_t func,uint8_t action){
         HID_Set_Action_Func(usb_handle,func,action);
     });
